Activar y descativar pagina

+++ public/js/index.js
@@ -1,4 +1,8 @@
 //Definicion de variables
+/* check pagina abierta */
+const checkPaginaAbierta = document.querySelector(".checkPaginaAbierta");
+const urlEstatusPagina = "http://localhost:3000/api/estatusPagina"
+
 const urlApi = "http://localhost:3000/api/restaurantes/";
 const contenedorRegistros = document.querySelector(".contenedorRegistros");
 /* const formNuevoRest = document.querySelector("#formNuevoRest"); */
@@ -17,6 +21,64 @@ const btnGuardarNuevoRest = document.getElementById("btnGuardarNuevoRest");
 const inputBuscador = document.getElementById("inputBuscador");
 const btnBuscador = document.getElementById("btnBuscador"); 
 
+/* Estatus de pagina*/
+fetch(urlEstatusPagina)
+.then(res => res.json())
+.then(data => {
+    let estatus = data[0].estatus;
+    if(estatus === "Activa"){
+        checkPaginaAbierta.children[0].checked = true;
+    }
+})
+
+const habilitarPagina = (estatus) => {
+    fetch(urlEstatusPagina, {
+        method:"PUT",
+        headers:{
+            "Content-Type":"application/json" 
+        },
+        body:JSON.stringify({
+            estatus:estatus
+        })
+    })
+}
+
+checkPaginaAbierta.onchange = () => {
+    let activa = checkPaginaAbierta.children[0].checked;
+    if(activa){
+        habilitarPagina("Activa");
+    }else{
+        Swal.fire({
+            title: 'Quieres Desactivar la pagina?',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Desactivar',
+            cancelButtonText: 'Cancelar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire({
+                    icon: 'question',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Terminamos por hoy',
+                    cancelButtonText: 'Tomar un pequeño descanso'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        habilitarPagina("Inactiva");
+                    }else{
+                        habilitarPagina("Descanso");
+                    }
+                })
+            }else{
+                checkPaginaAbierta.children[0].checked = true;
+            }
+        })
+    }
+}
+
 /* Limpiar inputs modal nuevo rest */
 btnNuevoRest.addEventListener("click", ()=>{
     imgNuevoRest.src = "./resources/img/icons/uploadImg.png";
@@ -28,42 +90,67 @@ btnNuevoRest.addEventListener("click", ()=>{
     limpiarChecks(checkDiaNuevoRest);
 });
 
-//Mostrar registros con fectth
-fetch(urlApi)
+//Mostrar registros Restaurantes con fecth
+const fetchRestaurantes = () =>{
+    fetch(urlApi)
     .then( res => res.json())
     .then(data => {
         data.forEach(restaurante => {
             baseDeDatos.push(restaurante);
         })
-        
-        renderizar(baseDeDatos);
+        console.log(baseDeDatos)
+        renderizarRestaurantes(baseDeDatos);
     })
     .catch(e => console.log(e))
+}
 
-    
+fetchRestaurantes();
+
 //Eliminar Restaurante 
 on(document, "click", ".btnBorrarRest", e => {
     const fila = e.target.parentNode.parentNode.parentNode;
     const nomRest = fila.children[2].innerHTML;
     const idRest = fila.firstElementChild.innerHTML;
-
-    Swal.fire({
-        title: `seguro quieres eliminar el registro de ${nomRest}?`,
-        showDenyButton: true,
-        confirmButtonText: 'Borrar',
-        denyButtonText: `Conservar`,
-    }).then((result) => {
-        if (result.isConfirmed) {
+    fetch(urlApi+idRest+"/menu")
+    .then(res => res.json())
+    .then(data => {
+        if(data.length != 0){
             Swal.fire({
-                showConfirmButton: false,
-                text:'El registro fue eliminado con exito',
-                })
-            fetch(urlApi+idRest,{
-                method: "DELETE"
+                icon: 'error',
+                title: 'Oops...',
+                text: 'No puedes Eliminar este Registro. Aun tiene categorias de menu vinculadas',
             })
-            .then(res => res.json())
-            .then(()=> location.reload())
-        } else if (result.isDenied) {
+        }else{
+            Swal.fire({
+                title: `Eliminar registro de <strong> ${nomRest}? </strong>`,
+                text: "No pueder revertir esta acción",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#dc3545',
+                cancelButtonColor: '#6c757d',
+                confirmButtonText: 'Borrar!'
+                }).then((result) => {
+                if (result.isConfirmed) {
+                    fetch(urlApi+idRest,{
+                        method: "DELETE"
+                    })
+                    .then(res => res.json())
+                    .then(()=>{
+                        Swal.fire({
+                            title:'Borrado!',
+                            text:'El regristro se borro con exito',
+                            icon:'success',
+                            timer:1500
+                        })
+                        baseDeDatos = [];
+                        fetchRestaurantes();
+                        prueba = baseDeDatos;
+                    } 
+                    )
+                    
+                }else if (result.isDenied) {
+                }
+                })
         }
     })
 })
@@ -92,21 +179,19 @@ const btnSiguiente = document.querySelector("#btnSiguiente");
 const informacionPagina = document.querySelector("#informacion-pagina");
 const elementosPorPagina = 2;
 let paginaActual = 1;
-const baseDeDatos = [];
+let baseDeDatos = [];
 let prueba = baseDeDatos;
-/* 
 
-*/
 const paginas = document.querySelector(".paginas");
 
 const avanzarPagina = () => {
     paginaActual = paginaActual + 1;
-    renderizar(prueba);
+    renderizarRestaurantes(prueba);
 }
 
 const retrocederPagina = () => {
     paginaActual = paginaActual - 1;
-	renderizar(prueba);
+	renderizarRestaurantes(prueba);
 }
 
 const obtenerRebanadaDeBaseDeDatos = (pagina = 1, datos) => {
@@ -134,7 +219,7 @@ function gestionarBotones(datos) {
 	}
 }
 
-const renderizar = (datos) => {
+const renderizarRestaurantes = (datos) => {
 	// Limpiamos los artículos anteriores del DOM
 	borrarElementosContenedor(".registroRestaurante", contenedorRegistros);
 	// Obtenemos los artículos paginados
